Added ChessGameFactory and MockChessGameFactory
