+++ src/lichess/MockChessGameFactory.py
@@ -1,0 +1,8 @@
+from lichess.IChessGameFactory import IChessGameFactory
+
+class MockChessGameFactory(IChessGameFactory):
+    def __init__(self):
+        pass
+
+    def create_game(self, game_info):
+        return {}
